% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fastmap.r
\name{cma}
\alias{cma}
\title{CMA}
\usage{
cma(x, k = 1)
}
\arguments{
\item{x}{The input matrix or dataframe.}

\item{k}{TODO (number of subspace iterations I think, don't remember)}
}
\description{
Coordinate Mapping Algorithm.
}

