% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cov.r
\name{cov2}
\alias{cov2}
\title{Covariance}
\usage{
cov2(x)
}
\arguments{
\item{x}{The input matrix, vector, or dataframe.}
}
\description{
An efficient method of quickly computing the covariance of
a matrix.  Like R's own \code{cov()}, this will cast dataframes
as matrices first, which is potentially a very memory expensive
operation.
}

