% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/pca.r
\name{pca}
\alias{pca}
\title{PCA}
\usage{
pca(x, retx = TRUE, center = TRUE, scale = FALSE, method = "svd")
}
\arguments{
\item{x}{The input matrix or dataframe.}

\item{retx}{Return the rotated variables?}

\item{center}{Center the matrix first?}

\item{scale}{Scale the matrix first?}

\item{method}{"svd" for svd of data matrix, or "eigcov" for eigenvalue decomposition
of the covariance matrix.}
}
\description{
Principal components analysis.
}

