% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pcapack-package.R
\docType{package}
\name{PCAPACK-package}
\alias{PCAPACK-package}
\title{PCAPACK}
\description{
PCAPACK is a set of utilities for performing principal components
analysis (including by means of truncated svd) very efficiently.
The package also contains efficient, related helper utilities
such as a replacement for R's covariance calculator.
}
\details{
\tabular{ll}{
   Package: \tab PCAPACK \cr
   Type: \tab Package \cr
   License: \tab MPL \cr
   Needs Compilation: \tab yes \cr
}
}
\author{
Drew Schmidt
}
\keyword{Package}

