% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/scale.r
\name{scale2}
\alias{scale2}
\title{Scale}
\usage{
scale2(x, center = TRUE, scale = TRUE)
}
\arguments{
\item{x}{The input matrix, vector, or dataframe.}

\item{center}{Logical; should the data be centered first? (yes!)}

\item{scale}{Logical; should the data be scaled first? (probably!)}
}
\description{
An efficient method of centering and/or scaling a matrix or vector.
Like R's own \code{scale()}, this will cast dataframes as matrices
first, which is potentially a very memory expensive operation.
}

