% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/svd.r
\name{svd2}
\alias{svd2}
\title{SVD}
\usage{
svd2(x, nu = min(n, p), nv = min(n, p))
}
\arguments{
\item{x}{The input matrix or dataframe.}

\item{nu}{Number of left singular vectors to return.}

\item{nv}{Number of right singular vectors to return.}
}
\description{
Singular value decomposition.
}

